Move kernel options stuff to options.{h,c}
